casted HP/UX warning on size_t to int
X-Trac-URL: https://ssl.icu-project.org/trac/ticket/535



git-svn-id: http://source.icu-project.org/repos/icu/icu/trunk@3726 0a8b766c-62c9-45f9-954d-7f2943c9ead0

+++ source/i18n/dbbi.cpp
@@ -274,18 +274,18 @@ BreakIterator *  DictionaryBasedBreakIterator::createBufferClone(void *stackBuff
     }
     if (text == NULL)
     {
-        bufferSizeNeeded = sizeof(DictionaryBasedBreakIterator);
+        bufferSizeNeeded = (int32_t) sizeof(DictionaryBasedBreakIterator);
         IterIsString = IterIsUChar = FALSE;
     }
     else if (text->getDynamicClassID() == StringCharacterIterator::getStaticClassID()) 
     {
-        bufferSizeNeeded = sizeof(struct bufferCloneStructString);
+        bufferSizeNeeded = (int32_t) sizeof(struct bufferCloneStructString);
         IterIsString = TRUE;
         IterIsUChar = FALSE;
     } 
     else if (text->getDynamicClassID() == UCharCharacterIterator::getStaticClassID()) 
     {
-        bufferSizeNeeded = sizeof(struct bufferCloneStructUChar);
+        bufferSizeNeeded = (int32_t) sizeof(struct bufferCloneStructUChar);
         IterIsString = FALSE;
         IterIsUChar = TRUE;
     }
